+++ lib/main.dart
@@ -4,6 +4,8 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:provider/provider.dart';
 import 'package:titan_app/pages/home/home_page.dart';
+import 'dart:async';
+
 import 'package:titan_app/pages/login/login_page.dart';
 import 'package:titan_app/pages/logs/logs_page.dart';
 import 'package:titan_app/pages/setting/setting_page.dart';
@@ -120,6 +122,9 @@ class _AppHomePageState extends State<AppHomePage> {
   }
 
   Widget _buildBottomNav(BuildContext context) {
+    var isLatestVersion =
+        Provider.of<VersionProvider>(context, listen: true).isLatestVersion;
+
     return Stack(
       children: [
         TabBottomBar(
@@ -127,8 +132,8 @@ class _AppHomePageState extends State<AppHomePage> {
           onItemTap: _onTapBottomNav,
           onItemLongTap: _onItemLongTap,
         ),
-        const Positioned(
-            right: 35, top: 16, child: UpdateRedPoint(isShow: true))
+        Positioned(
+            right: 35, top: 16, child: UpdateRedPoint(isShow: !isLatestVersion))
       ],
     );
   }
